# agri-frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AgriSmart Benin API",
  "description": "Backend Monolithique Modulaire pour le Challenge Agriculture Intelligente",
  "version": "4.0.0"
 },
 "paths": {
...
  "/api/v1/domains/{domain_id}/survey": {
   "put": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "Relevé de terrain de l'agent",
    "operationId": "save_survey_api_v1_domains__domain_id__survey_put",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "domain_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Domain Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SiteSurvey"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/domains/{domain_id}/orientation": {
   "put": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "Orientations de l'État pour cette terre",
    "operationId": "save_orientation_api_v1_domains__domain_id__orientation_put",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "domain_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Domain Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StateOrientation"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/domains/{domain_id}/photos": {
   "post": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "Ajouter une photo géolocalisée",
    "operationId": "add_photo_api_v1_domains__domain_id__photos_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "domain_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Domain Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_add_photo_api_v1_domains__domain_id__photos_post"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "List Photos",
    "operationId": "list_photos_api_v1_domains__domain_id__photos_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "domain_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Domain Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/domains/{domain_id}/photos/{photo_id}": {
   "get": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "Get Photo",
    "operationId": "get_photo_api_v1_domains__domain_id__photos__photo_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "domain_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Domain Id"
      }
     },
     {
      "name": "photo_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Photo Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "image/jpeg": {}
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/domains/{domain_id}/environment": {
   "post": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "Collecter sol, climat, relief et accessibilité",
    "operationId": "refresh_environment_api_v1_domains__domain_id__environment_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "domain_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Domain Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/domains/{domain_id}/readiness": {
   "get": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "Ce qui manque avant de générer le plan",
    "operationId": "readiness_api_v1_domains__domain_id__readiness_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "domain_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Domain Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/domains/{domain_id}/plans": {
   "post": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "Générer un plan de mise en valeur (IA)",
    "operationId": "generate_plan_api_v1_domains__domain_id__plans_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "domain_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Domain Id"
      }
     }
    ],
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlanOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "List Plans",
    "operationId": "list_plans_api_v1_domains__domain_id__plans_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "domain_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Domain Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/PlanOut"
         },
         "title": "Response List Plans Api V1 Domains  Domain Id  Plans Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/domains/plans/{plan_id}": {
   "get": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "Get Plan",
    "operationId": "get_plan_api_v1_domains_plans__plan_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "plan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Plan Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlanOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/domains/plans/{plan_id}/review": {
   "patch": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "Relecture par un expert (validation ou demande de révision)",
    "operationId": "review_plan_api_v1_domains_plans__plan_id__review_patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "plan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Plan Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PlanReview"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlanOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/domains/plans/{plan_id}/call-draft": {
   "get": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "Préremplir un appel à candidatures depuis le plan",
    "operationId": "call_draft_api_v1_domains_plans__plan_id__call_draft_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "plan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Plan Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/calls/{call_id}/applications/{application_id}/ai-review": {
   "post": {
    "tags": [
     "Domaine de l'État : préparation et plan IA"
    ],
    "summary": "Aide à l'analyse d'une candidature (IA)",
    "operationId": "review_application_api_v1_calls__call_id__applications__application_id__ai_review_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "call_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Call Id"
      }
     },
     {
      "name": "application_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Application Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "StateOrientation": {
    "properties": {
     "vocation": {
      "type": "string",
      "enum": [
       "vivrier",
       "rente_export",
       "semences",
       "jeunes_entrepreneurs",
       "mixte"
      ],
      "title": "Vocation"
     },
     "priority_crops": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 10,
      "title": "Priority Crops"
     },
     "excluded_crops": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 10,
      "title": "Excluded Crops"
     },
     "investment_level": {
      "type": "string",
      "enum": [
       "faible",
       "moyen",
       "eleve"
      ],
      "title": "Investment Level"
     },
     "mechanization": {
      "type": "string",
      "enum": [
       "manuelle",
       "attelee",
       "motorisee"
      ],
      "title": "Mechanization"
     },
     "min_valorization_pct": {
      "type": "number",
      "maximum": 100.0,
      "minimum": 10.0,
      "title": "Min Valorization Pct",
      "default": 80
     },
     "notes": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2000
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "vocation",
     "investment_level",
     "mechanization"
    ],
    "title": "StateOrientation"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}